requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0

# Face recognition
insightface>=0.7.0
onnxruntime>=1.12.0
# For GPU support (optional)
# onnxruntime-gpu>=1.12.0

# Additional ML libraries
scipy>=1.7.0
pandas>=1.3.0

# Development and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Optional: For advanced visualizations
plotly>=5.0.0
dash>=2.0.0

# Optional: For database support
sqlalchemy>=1.4.0